requirements/common.txt
dj-database-url==2.3.0
dj-static==0.0.6
Django==5.1
django-configurations==2.5.1
djangorestframework==3.15.2
ipython==8.29.0
openai==1.52.2
psycopg2-binary==2.9.10
